Assign users to tweets and fix user access

+++ polls/views.py
@@ -8,6 +8,7 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import UserCreationForm , AuthenticationForm
 from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 
 from polls.forms import TwittForm
 from datetime import datetime
@@ -23,19 +24,12 @@ from rest_framework_jwt.authentication import JSONWebTokenAuthentication
 
 
 class TweeterPage(APIView):
-    # def post(self , request ,format= None):
-        # serializer=TweetModelSerializer(data=request.data)
-        # if serializer.is_valid():
-            # serializer.save()
-            # return Response(serializer.data , status=status.HTTP_201_CREATED)
-        # return Response (serializer.errors , status =status.HTTP_400_BAD_REQUEST)
+    #@login_required
+ 
     authentication_classes = (JSONWebTokenAuthentication,)
     permission_classes = (IsAuthenticated,)
 
 
-class TweeterPage(APIView):
-    #@login_required
- 
     def get_object(self, id):
         try:
             return TweetModel.objects.get(id=id)
@@ -45,16 +39,30 @@ class TweeterPage(APIView):
     def post(self , request ,format= None):
         serializer=TweetModelSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            print("dakheele vlidated ")
+            #print(dir(serializer))
+            #serializer.validated_data['userr']=request.user
+            #print(serializer.validated_data)
+            #u = User.objects.filter(id=1).first()
+            print(request.user)
+            serializer.save(userr=request.user)
             return Response(serializer.data , status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status =status.HTTP_400_BAD_REQUEST)
    
     def delete(self, request, format=None):
         print(request.data['id'])
-        deleted = self.get_object(request.data['id'])
-       
-        deleted.delete()
-        return Response({"message": "tweet deleted"},status=status.HTTP_200_OK)
+        tweet_object = self.get_object(request.data['id'])
+        if request.user==tweet_object.userr:
+            deleted.delete()
+            return Response({"message": "tweet deleted"}, status=status.HTTP_200_OK)
+        else:
+            return Response(
+                {
+                    "message": "the access is denied ,"\
+                    "you cant delete other users' tweets "
+                },
+                status=status.HTTP_403_FORBIDDEN
+            )
 
     def get(self, request, format=None): 
         tweetmodels = TweetModel.objects.all()
